# internal/web/openapi.yaml
openapi: 3.0.3
info:
  title: Detent Stable JSON API
  version: 1.0.0
  description: |
    The stable, machine-readable JSON API for Detent operators and integrations.
    Dashboard HTML, HTMX fragments, and server-sent event streams are outside this
    catalog. Some cataloged operations also support an HTMX representation; this
    document describes only their non-HTMX JSON contract. All operations except
    discovery are registered only while Detent is in running mode.
servers:
  - url: /
...
paths:
...
  /api/v1/tracker/availability/clear:
    post:
      operationId: clearTrackerAvailability
      summary: Clear recorded tracker-availability conditions
      tags: [administration]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      x-detent-access: admin
      x-detent-scope: admin
      x-detent-echo-path: /api/v1/tracker/availability/clear
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                project_id: {type: string}
      responses:
        "200":
          description: Tracker availability records cleared.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Status"}
        "404": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Unavailable"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Detent static API token or scoped API key.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    githubSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature over the exact request body.
    intakeToken:
      type: apiKey
      in: header
      name: X-Detent-Intake-Token
    intakeBearer:
      type: http
      scheme: bearer
      description: Project intake secret, distinct from Detent API tokens and API keys.
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Problem"}
    Unavailable:
      description: A required orchestrator or backing service is degraded or unavailable.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
...
    APIError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
    CLIProblem:
      type: object
      description: RFC 9457-style Detent CLI JSON error envelope documented in docs/cli.md.
      required: [type, code, title, detail, exit_code]
      properties:
        type: {type: string, format: uri}
        code: {type: string}
        title: {type: string}
        detail: {type: string}
        exit_code: {type: integer}
        suggested_fix: {type: string}
        did_you_mean:
          type: array
          items: {type: string}
        docs_url: {type: string, format: uri}
    Problem:
      description: Stable Detent machine-error forms. HTTP API routes emit APIError; CLI JSON commands emit CLIProblem.
      oneOf:
        - $ref: "#/components/schemas/APIError"
        - $ref: "#/components/schemas/CLIProblem"
...
    Status:
      type: object
      additionalProperties: true
...